Move colorSuggestions type to types dir
