+++ src/components/layout/BaseColorsSection/ColorSuggestionsBox.tsx
@@ -7,7 +7,7 @@ import {
 import { IconButton, ListItem } from 'components/general';
 import { DoneIcon } from 'assets/icons';
 import { isHexColorLight } from 'utils';
-import type { ColorSuggestions } from 'constants/colorSuggestions';
+import type { ColorSuggestions } from 'types';
 
 const ColorSuggestionsBox = <T extends ColorSuggestions>({
   baseColor,
