+++ src/types/index.ts
@@ -1,1 +1,3 @@
+export type * from './codeGen';
+export type * from './colorSuggestions';
 export type * from './utils';
